move label handler to uplot plugins
