+++ plugin/blocks/src/components/graph/components/uplot-graph.js
@@ -52,112 +52,12 @@ export default function UplotGraph({
 		containerRef,
 		showLegend,
 		type,
-		orientation
-	}), [title, data, series, meta, containerRef, showLegend, type, orientation]);
+		orientation,
+		dimension
+	}), [title, data, series, meta, containerRef, showLegend, type, orientation, dimension]);
 
 	// Get graph options from the hook
 	const { data: d, ...options } = useGraphOptions(graphOptionsParams);
-
-	// Create a custom plugin for handling legend label click events only
-	const legendLabelClickPlugin = React.useMemo(() => {
-		return {
-			hooks: {
-				ready: [
-					(u) => {
-						// Function to handle legend label events
-						const handleLegendEvents = () => {
-							// Find all legend labels
-							const labels = u.root.querySelectorAll('.u-label');
-
-							// Add click event for specific dimensions
-							if (['atomic_site_id', 'http_host'].includes(dimension)) {
-								labels.forEach(label => {
-									// Skip labels that already have event handlers attached
-									if (label.dataset.hasClickEvent === 'true') {
-										return;
-									}
-
-									label.style.cursor = 'pointer';
-									const clickHandler = (e) => {
-										e.stopPropagation();
-										console.log(`Clicked label: ${label.textContent}`);
-
-										var site_url = label.textContent;
-										if (isNaN(site_url)) {
-											site_url = site_url.replace(/\./g, "-");
-										}
-										window.location.href = `/sites/${site_url}/metrics/`;
-									};
-
-									label.addEventListener('click', clickHandler);
-
-									// Mark this label as having click event handler attached
-									label.dataset.hasClickEvent = 'true';
-								});
-							}
-						};
-
-						// Initial setup with a delay to ensure DOM is ready
-						setTimeout(handleLegendEvents, 100);
-
-						// Set up a MutationObserver to watch for changes to the legend
-						const observer = new MutationObserver((mutations) => {
-							// Re-apply event handlers when the DOM changes
-							handleLegendEvents();
-						});
-
-						// Start observing the chart root for changes
-						observer.observe(u.root, {
-							childList: true,
-							subtree: true
-						});
-
-						// Store the observer for cleanup
-						u.legendObserver = observer;
-					}
-				],
-				destroy: [
-					(u) => {
-						// Clean up the observer when the chart is destroyed
-						if (u.legendObserver) {
-							u.legendObserver.disconnect();
-						}
-
-						// Remove data attributes from labels
-						const labels = u.root.querySelectorAll('.u-label');
-						labels.forEach(label => {
-							if (label.dataset.hasClickEvent === 'true') {
-								delete label.dataset.hasClickEvent;
-							}
-						});
-					}
-				]
-			}
-		};
-	}, [dimension]);
-
-	// Add the plugin to the options
-	React.useEffect(() => {
-		if (uplotInstanceRef.current) {
-			// Apply the plugin to the existing chart
-			const u = uplotInstanceRef.current;
-
-			// Clean up any existing observer
-			if (u.legendObserver) {
-				u.legendObserver.disconnect();
-			}
-
-			// Set up the plugin hooks
-			legendLabelClickPlugin.hooks.ready.forEach(hook => hook(u));
-
-			// Clean up when the component unmounts or when dependencies change
-			return () => {
-				if (u.legendObserver) {
-					u.legendObserver.disconnect();
-				}
-			};
-		}
-	}, [legendLabelClickPlugin, uplotInstanceRef.current]);
 
 	return (
 		<>
